+++ src/Broadway/ReadModel/ElasticSearch/ElasticSearchRepositoryFactory.php
@@ -34,6 +34,8 @@ class ElasticSearchRepositoryFactory implements RepositoryFactoryInterface
      */
     public function create($name, $class, array $notAnalyzedFields = array())
     {
-        return new ElasticSearchRepository($this->client, $this->serializer, $name, $class, $notAnalyzedFields);
+        $objectRepository = new ElsaticSearchSerializableObjectRepository($this->client, $this->serializer, $name, $class, $notAnalyzedFields);
+
+        return new ElasticSearchRepository($objectRepository);
     }
 }
